Split BPF E2E tests Some tests that are part of the fedora-e2e test suite are intermittendly failing (a.k.a. flaking). Given that the whole suite takes up to 40 minutes to run. This allows for auto retry of specific tests rather than the entire suite.
