+++ test/e2e_noflake_test.go
@@ -1,0 +1,167 @@
+//go:build !flake
+// +build !flake
+
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package e2e_test
+
+func (e *e2e) runClusterWideTests(nodes []string) {
+	// Execute the test cases. Each test case should cleanup on its own and
+	// leave a working operator behind.
+	e.logf("testing cluster-wide operator")
+	testCases := []struct {
+		description string
+		fn          func(nodes []string)
+	}{
+		{
+			"Seccomp: Verify default and example profiles",
+			e.testCaseDefaultAndExampleProfiles,
+		},
+		{
+			"Seccomp: Run a test pod",
+			e.testCaseRunPod,
+		},
+		{
+			"Seccomp: Verify base profile merge",
+			e.testCaseBaseProfile,
+		},
+		{
+			"Seccomp: Delete profiles",
+			e.testCaseDeleteProfiles,
+		},
+		{
+			"Seccomp: Re-deploy the operator",
+			e.testCaseReDeployOperator,
+		},
+		{
+			"Log Enricher",
+			e.testCaseLogEnricher,
+		},
+		{
+			"SELinux: sanity check",
+			e.testCaseSelinuxSanityCheck,
+		},
+		{
+			"SELinux: base case (install policy, run pod and delete)",
+			e.testCaseSelinuxBaseUsage,
+		},
+		{
+			"SELinux: Metrics (update, delete)",
+			e.testCaseSelinuxMetrics,
+		},
+		{
+			"SPOD: Update SELinux flag",
+			e.testCaseSPODUpdateSelinux,
+		},
+		{
+			"SPOD: Change verbosity",
+			e.testCaseVerbosityChange,
+		},
+		{
+			"Seccomp: make sure statuses for profiles with long names can be listed",
+			e.testCaseLongSeccompProfileName,
+		},
+	}
+	for _, testCase := range testCases {
+		tc := testCase
+		e.Run("cluster-wide: "+tc.description, func() {
+			tc.fn(nodes)
+		})
+	}
+
+	e.Run("cluster-wide: Seccomp: Verify profile recording hook", func() {
+		e.testCaseProfileRecordingStaticPodHook()
+		e.testCaseProfileRecordingKubectlRunHook()
+		e.testCaseProfileRecordingMultiContainerHook()
+		e.testCaseProfileRecordingDeploymentHook()
+	})
+
+	e.Run("cluster-wide: Seccomp: Verify profile recording logs", func() {
+		e.testCaseProfileRecordingStaticPodLogs()
+		e.testCaseProfileRecordingMultiContainerLogs()
+		e.testCaseProfileRecordingDeploymentLogs()
+	})
+
+	e.Run("cluster-wide: Selinux: Verify SELinux profile recording logs", func() {
+		e.testCaseProfileRecordingStaticPodSELinuxLogs()
+		e.testCaseProfileRecordingMultiContainerSELinuxLogs()
+		e.testCaseProfileRecordingSelinuxDeploymentLogs()
+	})
+}
+
+func (e *e2e) runNamespacedTests(nodes []string) {
+	testCases := []struct {
+		description string
+		fn          func(nodes []string)
+	}{
+		{
+			"Seccomp: Verify default and example profiles",
+			e.testCaseDefaultAndExampleProfiles,
+		},
+		{
+			"Seccomp: Run a test pod",
+			e.testCaseRunPod,
+		},
+		{
+			"Seccomp: Verify base profile merge",
+			e.testCaseBaseProfile,
+		},
+		{
+			"Seccomp: Delete profiles",
+			e.testCaseDeleteProfiles,
+		},
+		{
+			"Seccomp: Metrics",
+			e.testCaseSeccompMetrics,
+		},
+		{
+			"Seccomp: Re-deploy the operator",
+			e.testCaseReDeployNamespaceOperator,
+		},
+		{
+			"Log Enricher",
+			e.testCaseLogEnricher,
+		},
+		{
+			"SELinux: sanity check",
+			e.testCaseSelinuxSanityCheck,
+		},
+		{
+			"SELinux: base case (install policy, run pod and delete)",
+			e.testCaseSelinuxBaseUsage,
+		},
+		{
+			"SPOD: Update SELinux flag",
+			e.testCaseSPODUpdateSelinux,
+		},
+		{
+			"SPOD: Change verbosity",
+			e.testCaseVerbosityChange,
+		},
+		{
+			"Seccomp: make sure statuses for profiles with long names can be listed",
+			e.testCaseLongSeccompProfileName,
+		},
+	}
+
+	for _, testCase := range testCases {
+		tc := testCase
+		e.Run("namespaced: "+tc.description, func() {
+			tc.fn(nodes)
+		})
+	}
+}
